cxl: export internal structs for external Type2 drivers
In preparation for type2 support, move structs and functions a type2
driver will need to access to into a new shared header file.

Differentiate between public and private data to be preserved by type2
drivers.

Tested-by: Alison Schofield <alison.schofield@intel.com>
Link: https://patch.msgid.link/20260306164741.3796372-3-alejandro.lucero-palau@amd.com
(cherry picked from commit 005869886d1d370afb6c10cd40709d956960e9c2)
